% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethyExpressCor.R
\name{MethyExpressCor}
\alias{MethyExpressCor}
\title{Correlation analysis of methylation and expression}
\usage{
MethyExpressCor(methydata, expressdata, pairsdata, method = "spearman")
}
\arguments{
\item{methydata}{Mehtylation data matrix}

\item{expressdata}{Expression data matrix}

\item{pairsdata}{The gene and cpg pairs for correlation analysis, column V1 is the cpg probe and column V2 is the gene name}

\item{method}{The method used to calculate correlation Defaults to \pkg{spearman}}
}
\value{
GCPs correlation statistcis
}
\description{
This function perform correlation analysis of DNA methylation and gene expression between gene and cpg pairs (GCPs).
}
\examples{
MethyExpressCor(methydata, expressdata, pairsdata)
}
\author{
WKL
}
\keyword{GCPs,}
\keyword{correlation}
